# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NadlanConnect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /favorites/{listingId}:
    post:
      tags: [Favorites]
      operationId: addFavorite
      summary: Add a listing to favorites
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
    delete:
      tags: [Favorites]
      operationId: removeFavorite
      summary: Remove a listing from favorites
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'

  # ── Leads ─────────────────────────────────────────────────────────────────
  /leads:
    get:
      tags: [Leads]
      operationId: listLeads
      summary: List leads (buyer sees own, pro sees received, admin sees all)
      responses:
        '200':
          description: Leads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
    post:
      tags: [Leads]
      operationId: createLead
      summary: Send a contact request (buyer contacts a listing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadInput'
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
  /leads/{leadId}:
    get:
      tags: [Leads]
      operationId: getLead
      summary: Get a single lead with messages
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lead detail with messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadDetail'
    patch:
      tags: [Leads]
      operationId: updateLeadStatus
      summary: Update lead status (pro only)
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadStatusUpdate'
      responses:
        '200':
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'

  # ── Messages ──────────────────────────────────────────────────────────────
  /leads/{leadId}/messages:
    get:
      tags: [Messages]
      operationId: getLeadMessages
      summary: Get messages for a lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      tags: [Messages]
      operationId: sendMessage
      summary: Send a message in a lead thread
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Favorite:
      type: object
      required: [id, userId, listingId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        listingId:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── Leads ──
    Lead:
      type: object
      required: [id, listingId, buyerId, message, status, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        listingTitle:
          type: [string, 'null']
        buyerId:
          type: string
        buyerName:
          type: [string, 'null']
        buyerEmail:
          type: [string, 'null']
        message:
          type: string
        status:
          type: string
          enum: [new, contacted, closed]
        createdAt:
          type: string
          format: date-time
    LeadDetail:
      type: object
      required: [lead, messages]
      properties:
        lead:
          $ref: '#/components/schemas/Lead'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    LeadInput:
      type: object
      required: [listingId, message]
      properties:
        listingId:
          type: integer
        message:
          type: string
          minLength: 1
    LeadStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, contacted, closed]

    # ── Messages ──
    Message:
      type: object
      required: [id, leadId, senderId, body, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        senderId:
          type: string
        senderName:
          type: [string, 'null']
        senderAvatar:
          type: [string, 'null']
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1

    # ── Dashboard ──
...
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...