# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface Favorite {
  id: number;
  userId: string;
  listingId: number;
  createdAt: Date;
}