# lib/api-zod/src/generated/types/leadDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from './lead';
import type { Message } from './message';

export interface LeadDetail {
  lead: Lead;
  messages: Message[];
}