# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  listingId: number;
  /** @minLength 1 */
  message: string;
}