# lib/api-zod/src/generated/types/leadStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatusUpdateStatus } from './leadStatusUpdateStatus';

export interface LeadStatusUpdate {
  status: LeadStatusUpdateStatus;
}